updated to suppose device suspension/sleep mode
When a device goes into suspension or sleep mode, javascript stops running which will prevent the timer from running. When the device was "awoken", even long after the session timer was done, the warning dialog would pop up with a timer showing the "perceived" time left, even though the actual session has already timed out on the server. This change now stores the loaded time of the page and checks the current time difference from the loaded time to determine if the session has expired and redirect you immediately.